refactor(#180): split hook into upload, restore, and orchestrator files
